sync database just on dev

+++ models/models.js
@@ -1,3 +1,5 @@
+require('dotenv').config();
+
 
 const dbConfig = require("../config/database.js");
 
@@ -99,18 +101,19 @@ db.updateOrCreate = async (model, where, newItem) => {
         return model
           .create(newItem)
           .then((item) => {
-            return  item;
+            return item;
           })
       }
       // Found an item, update it
       return model
         .update(newItem, { where: where })
         .then((item) => {
-          return  item
+          return item
         });
     })
-  }
-
-db.sequelize.sync({ force: false });
+}
+if (process.env.ENV === 'DEV') {
+  db.sequelize.sync({ force: false });
+}
 
 module.exports = db;
